🔥[MODIFY][backend][이홍열] 동네정보 게시물 삭제기능 수정  - 게시물 삭제시 작성한 유저만 삭제할수 있도록 수정  - 게시물 작성 한 유저 외 게시물 삭제 시 에러메시지 발생  - app.js 서버에러 소스코드 삭제

+++ backend/services/cmpostService.js
@@ -32,7 +32,7 @@ const updateCmpost = async (
   );
 };
 
-const deleteCmpost = async (postId) => {
+const deleteCmpost = async (postId, userId) => {
   const checkPostId = await cmpostDao.checkCmpostId(postId);
 
   if (!checkPostId) {
@@ -40,6 +40,14 @@ const deleteCmpost = async (postId) => {
     error.statusCode = 400;
     throw error;
   }
+
+  const getUserIdByCmpostId = await cmpostDao.getUserIdByCmpostId(postId);
+  if (userId !== getUserIdByCmpostId) {
+    const error = new Error('NOT_AUTHORIZED_USER');
+    error.statusCode = 403;
+    throw error;
+  }
+
   return cmpostDao.deleteCmpost(postId);
 };
